Refactored common scraper code
